refactor names and remove unused components
